+++ Common/Actions/CityAction.cs
@@ -11,7 +11,7 @@ namespace Common.Actions
     {
         public int IntersectionIndex { get; init; }
 
-        public CityAction(int playerIdx, int intersectionIndex)
+        public CityAction(sbyte playerIdx, int intersectionIndex)
             : base(playerIdx)
         {
             IntersectionIndex = intersectionIndex;
@@ -102,7 +102,7 @@ namespace Common.Actions
             return hasSettlement && ownsSettlement && canAfford;
         }
 
-        public static List<Action> GetActionsForState(GameState state, int playerIdx)
+        public static List<Action> GetActionsForState(GameState state, sbyte playerIdx)
         {
             List<Action> actions = new List<Action>();
