+++ Common/Actions/FourToOneTradeAction.cs
@@ -11,7 +11,7 @@ namespace Common.Actions
         public ResourceCardType InputType { get; init; }
         public ResourceCardType OutputType { get; init; }
 
-        public FourToOneTradeAction(int playerIdx, ResourceCardType inputType, ResourceCardType outputType)
+        public FourToOneTradeAction(sbyte playerIdx, ResourceCardType inputType, ResourceCardType outputType)
             : base(playerIdx)
         {
             InputType = inputType;
@@ -73,7 +73,7 @@ namespace Common.Actions
             return cardTypesDiffer && playerHasInput && bankHasOutput;
         }
 
-        public static List<Action> GetActionsForState(GameState state, int playerIdx)
+        public static List<Action> GetActionsForState(GameState state, sbyte playerIdx)
         {
             List<Action> actions = [];
