+++ Common/Actions/SecondInitialSettlementAction.cs
@@ -16,7 +16,7 @@ namespace Common.Actions
 
         public int IntersectionIndex { get; init; }
 
-        public SecondInitialSettlementAction(int playerIdx, int intersectionIndex)
+        public SecondInitialSettlementAction(sbyte playerIdx, int intersectionIndex)
             : base(playerIdx)
         {
             IntersectionIndex = intersectionIndex;
@@ -129,7 +129,7 @@ namespace Common.Actions
             History = null;
         }
 
-        public static List<Action> GetActionsForState(GameState state, int playerIdx)
+        public static List<Action> GetActionsForState(GameState state, sbyte playerIdx)
         {
             List<Action> actions = new List<Action>();
